+++ src/ui-kit/hierarchical/autocomplete/autocomplete.component.scss
