+++ src/ui-kit/hierarchical/hierarchical-tree-grid/hierarchical-tree-grid.component.ts
@@ -3,7 +3,7 @@ import {
   Output, EventEmitter, ChangeDetectorRef
 } from '@angular/core';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { SamSortDirective } from '../../../components'
+import { SamSortDirective } from '../../components'
 import { HierarchicalDataSource } from './data-source';
 import { SamHierarchicalTreeGridConfiguration } from '../models/SamHierarchicalTreeGridConfiguration';
