+++ src/ui-kit/hierarchical/hierarchical-tree-header/hierarchical-tree-header.component.scss
