+++ src/ui-kit/hierarchical/hierarchical-tree/hierarchical-tree.component.scss
