+++ src/ui-kit/hierarchical/hierarchical-tree/hierarchical-tree.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, Input } from "@angular/core";
 import { Observable, BehaviorSubject } from "rxjs";
 import { SamHiercarchicalServiceInterface, SamHiercarchicalServiceResult } from "../hierarchical-interface";
 import { SamHierarchicalTreeConfiguration } from "../models/SamHierarchicalTreeConfiguration";
-import { Sort } from "../../../components/data-table/sort.directive";
+import { Sort } from "../../components/data-table/sort.directive";
 import { first } from 'rxjs/operators';
 
 @Component({
