+++ src/ui-kit/hierarchical/hierarchical/hierarchical.component.scss
