+++ src/ui-kit/hierarchical/models/SamHierarchicalAutocompleteConfiguration.ts
