+++ src/ui-kit/hierarchical/models/SamHierarchicalConfiguration.ts
