+++ src/ui-kit/hierarchical/models/SamHierarchicalSelectedResultConfiguration.ts
