+++ src/ui-kit/hierarchical/models/SamHierarchicalTreeConfiguration.ts
