+++ src/ui-kit/hierarchical/models/SamHierarchicalTreeHeaderConfiguration.ts
