+++ src/ui-kit/hierarchical/selected-result/selected-result.component.scss
