+++ src/ui-kit/hierarchical/selected-result/selected-result.component.spec.ts
