+++ src/ui-kit/hierarchical/selected-result/selected-result.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, TemplateRef, forwardRef } from '@angular/core';
 import { HierarchicalTreeSelectedItemModel } from '../hierarchical-tree-selectedItem.model';
 import { SelectedResultConfiguration } from '../models/SamHierarchicalSelectedResultConfiguration';
-import { NG_VALUE_ACCESSOR, ControlValueAccessor, FormControl } from '@angular/forms';
+import { NG_VALUE_ACCESSOR, ControlValueAccessor } from '@angular/forms';
 const Hierarchical_SelectedResult_VALUE_ACCESSOR: any = {
   provide: NG_VALUE_ACCESSOR,
   useExisting: forwardRef(() => SamHierarchicalSelectedResultComponent),
